requirements.txt
python-dotenv==1.0.0
sqlalchemy==2.0.23
requests==2.31.0
fastapi==0.109.0
uvicorn==0.27.0
streamlit==1.35.0
playwright==1.41.0
apscheduler==3.10.4
transformers==4.36.0
torch==2.1.0
plotly==5.18.0
pyyaml==6.0.1
pytest==8.0.0
